requirements.txt
# python utils
Pillow==7.1.0
requests==2.22.0
deprecated==1.2.6
beautifulsoup4==4.8.0
openpyxl==3.0.0
python-dateutil==2.8.0

Django==2.2.13

# django utils
django_widget_tweaks==1.4.5
django-bootstrap-form==3.4
django-sql-explorer==1.1.3
dj_database_url==0.5.0
django-storages==1.7.2
pik-django-utils==2.0.1

# HISTORY
django-simple-history==2.7.3

# API
djangorestframework==3.10.3
djangorestframework-filters==1.0.0.dev0
django-filter==2.2.0
django-crispy-forms==1.7.2
markdown==3.1.1
# API SCHEMA
drf-yasg==1.16.1

# db
redis==3.3.8
psycopg2==2.8.3
django-redis==4.10.0

# celery
celery==4.3.1
kombu<5.0
django_celery_results==1.1.2
celery-redbeat==0.13.0

# sentry
sentry-sdk==0.14.1

# prod
uwsgi==2.0.18

# dev
django_debug_toolbar==2.2
django_extensions==2.2.1
werkzeug==0.16.0

# tests
factory_boy==2.12.0
freezegun==0.3.12
pdbpp==0.10.0
pytest==5.2.0
pytest-mock==1.11.0
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-selenium==1.17.0
pytest-base-url==1.4.1
pytest-benchmark==3.2.2
pytest-html==2.0.0
# tests/style
prospector==1.1.2
isort<5.0.0 
pylint==2.1.1
flake8<3.6.0
pep8-naming==0.8.2

# OIDC
django_cors_headers==2.5.2
django-oauth-toolkit<1.1.0
django-rest-framework-social-oauth2==1.1.0
pyjwkest==1.4.0
future>=0.16.0,<0.17.0
social_auth_core>=1.7.0,<2.0.0

# lib/codegen
prance==0.16.1
openapi-spec-validator==0.2.8

# metrics/monitoring
ddtrace==0.29.0
datadog==0.30.0
elastic-apm==5.1.2
django-health-check==3.11.0
